backend/requirements.txt
fastapi
pydantic
pandas
RPi.GPIO
adafruit-circuitpython-htu21d
SQLAlchemy
uvicorn
plotly
syncer
dash
fastapi-utils